+++ Changelog.md
@@ -54,6 +54,7 @@
 * Do not hangup on parsing Coordinates like `Donut!A7:A7,Donut!A16:A16`
 * Do not hangup on parsing Coordinates with `#`
 * Fix crash on parsing Coordinates consist of `!`
+* Fix crash on `Delimeter`
 
 ## 0.2.0 (2017-03-25)
 ### New features